+++ packages/iov-keycontrol/types/keyring-entries/ed25519.d.ts
@@ -1,4 +1,5 @@
 import { PrehashType, SignableBytes } from "@iov/bcp-types";
+import { Ed25519Keypair } from "@iov/crypto";
 import { ChainId, SignatureBytes } from "@iov/tendermint-types";
 import { KeyringEntry, KeyringEntryId, KeyringEntryImplementationIdString, KeyringEntrySerializationString, LocalIdentity, PublicIdentity } from "../keyring";
 import { ValueAndUpdates } from "../valueandupdates";
@@ -16,7 +17,7 @@ export declare class Ed25519KeyringEntry implements KeyringEntry {
     private readonly labelProducer;
     constructor(data?: KeyringEntrySerializationString);
     setLabel(label: string | undefined): void;
-    createIdentity(): Promise<LocalIdentity>;
+    createIdentity(keypair?: Ed25519Keypair): Promise<LocalIdentity>;
     setIdentityLabel(identity: PublicIdentity, label: string | undefined): void;
     getIdentities(): ReadonlyArray<LocalIdentity>;
     createTransactionSignature(identity: PublicIdentity, transactionBytes: SignableBytes, prehashType: PrehashType, _: ChainId): Promise<SignatureBytes>;